requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
asyncpg
pandas
sqlalchemy
langchain
langgraph
google-generativeai
chromadb
openai
supermemory
wikipedia-api

